se agregan documentacion y pruebas unitarias

+++ UsuarioMicroServicio/src/main/java/cl/duoc/UsuarioMicroServicio/entity/Usuario.java
@@ -14,7 +14,7 @@ public class Usuario {
 
     @Id
     @Column(name = "ID_USUARIO")
-    private Long idUsuario;
+    private int idUsuario;
 
     @Column(name = "NOMBRE")
     private String nombre;
